Remove Providers folder in Server project
