+++ ArmaForces.Arma.Server.Tests/Providers/Parameters/ServerParametersProviderTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 using ArmaForces.Arma.Server.Config;
-using ArmaForces.Arma.Server.Providers.Parameters;
+using ArmaForces.Arma.Server.Features.Parameters.Providers;
 using ArmaForces.Arma.Server.Tests.Helpers;
 using AutoFixture;
 using FluentAssertions;
