+++ ArmaForces.Arma.Server/Config/RegistryReader.cs
@@ -1,10 +1,15 @@
-﻿using Microsoft.Win32;
+﻿using System;
+using Microsoft.Win32;
 
-namespace ArmaForces.Arma.Server.Config {
-    public class RegistryReader : IRegistryReader {
+namespace ArmaForces.Arma.Server.Config
+{
+    internal class RegistryReader : IRegistryReader
+    {
         public object? GetValueFromLocalMachine(string subKey, string value)
-            => Registry.LocalMachine
-                .OpenSubKey(subKey)
-                ?.GetValue(value);
+            => OperatingSystem.IsWindows()
+                ? Registry.LocalMachine
+                    .OpenSubKey(subKey)?
+                    .GetValue(value)
+                : null;
     }
 }
