+++ ArmaForces.Arma.Server/Features/Parameters/Providers/IParametersProvider.cs
@@ -1,6 +1,4 @@
-﻿using ArmaForces.Arma.Server.Features.Parameters;
-
-namespace ArmaForces.Arma.Server.Providers.Parameters
+﻿namespace ArmaForces.Arma.Server.Features.Parameters.Providers
 {
     public interface IParametersProvider
     {
