+++ ArmaForces.Arma.Server/Features/Processes/ArmaProcessDiscoverer.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Extensions;
 using ArmaForces.Arma.Server.Features.Parameters;
+using ArmaForces.Arma.Server.Features.Parameters.Extractors;
 using CSharpFunctionalExtensions;
 using Microsoft.Extensions.Logging;
